refactor: use a single axios instance for create and capture order
